+++ examples/protocol_test/src/bin/sender.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
 use fe2o3_amqp::sasl_profile::SaslProfile;
-use fe2o3_amqp::types::messaging::message::BodySection;
+use fe2o3_amqp::types::messaging::message::Body;
 use fe2o3_amqp::types::messaging::Message;
 use fe2o3_amqp::Connection;
 use fe2o3_amqp::Sendable;
@@ -24,15 +24,15 @@ async fn main() {
 
     tracing::subscriber::set_global_default(subscriber).expect("setting default subscriber failed");
 
-    let addr = "localhost:5671";
-    let domain = "localhost";
-    let stream = TcpStream::connect(addr).await.unwrap();
-    let connector = native_tls::TlsConnector::builder()
-        .danger_accept_invalid_certs(true)
-        .build()
-        .unwrap();
-    let connector = tokio_native_tls::TlsConnector::from(connector);
-    let tls_stream = connector.connect(domain, stream).await.unwrap();
+    // let addr = "localhost:5671";
+    // let domain = "localhost";
+    // let stream = TcpStream::connect(addr).await.unwrap();
+    // let connector = native_tls::TlsConnector::builder()
+    //     .danger_accept_invalid_certs(true)
+    //     .build()
+    //     .unwrap();
+    // let connector = tokio_native_tls::TlsConnector::from(connector);
+    // let tls_stream = connector.connect(domain, stream).await.unwrap();
 
     // let mut connection = Connection::open("connection-1", "amqp://guest:guest@localhost:5671")
     //     .await
@@ -43,13 +43,13 @@ async fn main() {
         .max_frame_size(1000)
         .channel_max(9)
         .idle_time_out(50_000 as u32)
-        .sasl_profile(SaslProfile::Plain {
-            username: "guest".into(),
-            password: "guest".into(),
-        })
-        .open_with_stream(tls_stream)
+        // .sasl_profile(SaslProfile::Plain {
+        //     username: "guest".into(),
+        //     password: "guest".into(),
+        // })
+        // .open_with_stream(tls_stream)
         // .open("amqp://localhost:5672")
-        // .open("amqp://guest:guest@localhost:5672")
+        .open("amqp://guest:guest@localhost:5672")
         .await
         .unwrap();
 
@@ -72,7 +72,7 @@ async fn main() {
     //     .await
     //     .unwrap();
 
-    let body = BodySection::from("hello");
+    let body = Body::from("hello");
     // let message = Message::from("hello");
     let message = Message::from(body);
     let message = Sendable::from(message);
@@ -101,7 +101,7 @@ async fn main() {
     // let result = fut.await;
     // println!("fut {:?}", result);
 
-    sender.close().await.unwrap();
+    sender.detach().await.unwrap();
 
     // let receiver = Receiver::attach(&mut session, "rust-receiver-link-1", "q1")
     //     .await